% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{addStrings}
\alias{addStrings}
\title{Add two strings alphabetically.}
\arguments{
\item{s1:}{string}

\item{s2:}{string}
}
\value{
string which contains given strings
}
\description{
Add two strings alphabetically.
}
\examples{
addStrings("abc", "def") => "abcdef"
addStrings("def", "abc") => "abcdef"
}
