% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{addTwoVectors}
\alias{addTwoVectors}
\title{Add two CharacterVectors.}
\arguments{
\item{v1:}{CharacterVector}

\item{v2:}{CharacterVector}
}
\value{
CharacterVector that has all the elements of the given vectors. Elements in v1 come first.
}
\description{
Add two CharacterVectors.
}
\examples{
addTwoVectors(c("A", "B", "C"), c("D", "E", "F")) => c("A", "B", "C", "D", "E", "F")
addTwoVectors(c("D", "E", "F"), c("A", "B", "C")) => c("D", "E", "F", "A", "B", "C")
}
