% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bin_to_img}
\alias{bin_to_img}
\title{Converts binary code to an image file.}
\usage{
bin_to_img(bin, name)
}
\arguments{
\item{bin:}{binary code of image}

\item{name:}{name of the created file}
}
\description{
Converts binary code to an image file.
}
