% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clusters_with_ids}
\alias{clusters_with_ids}
\title{Get clusters with patients IDs.}
\usage{
clusters_with_ids(patients_ids, clusters_with_indexes)
}
\arguments{
\item{patients_ids:}{numeric vector}

\item{clusters_with_indexes:}{numeric vector}
}
\value{
vector which has a list for every cluster. Lists contain patients IDs, who belong to the cluster.
}
\description{
Two vectors given as arguments have to have the same length.
Every element in the second argument shows into which cluster a patient in
the first argument on the same index belongs.
}
