% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{collectPatientData}
\alias{collectPatientData}
\title{Collect data about patients.}
\usage{
collectPatientData(data, IDs, n)
}
\arguments{
\item{data:}{dataframe}

\item{IDs:}{numeric vector}

\item{n:}{integer, n > 0 - length of drug trajectory}
}
\value{
list with vectors named "IDs" and "drugs". "IDs" hold information about patients IDs and "drugs" contains n-length drug trajectory per patient
}
\description{
Finds drug trajectory for every patient.
}
