% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compare_trajectories}
\alias{compare_trajectories}
\title{Compares all patients drug trajectories with each other}
\usage{
compare_trajectories(patients_data, punishments, dtw = FALSE)
}
\arguments{
\item{patient_data:}{list with patients IDs and trajectories}

\item{punishments:}{numeric vector}

\item{dtw:}{boolean if we want to use DTW algorithm}
}
\value{
symmetrical matrix which values are distances between patients trajectories
}
\description{
Compares all patients drug trajectories with each other
}
