% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{count_all_patient_drugs}
\alias{count_all_patient_drugs}
\title{Count occurrences of all drugs in patients records.}
\usage{
count_all_patient_drugs(data, id)
}
\arguments{
\item{data:}{dataframe}

\item{id:}{integer}
}
\value{
dataframe which shows count of all different drugs written to patient
}
\description{
Count occurrences of all drugs in patients records.
}
