% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countDrugsInCluster}
\alias{countDrugsInCluster}
\title{Find the occurrences of given drugs per patient.}
\usage{
countDrugsInCluster(data, IDs, drugs)
}
\arguments{
\item{data:}{dataframe}

\item{IDs:}{numeric vector}

\item{drugs:}{character vector}
}
\value{
numeric matrix which has count of given drugs per patient
}
\description{
Find the occurrences of given drugs per patient.
}
