% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{distance_between_codes}
\alias{distance_between_codes}
\title{Calculate distance between two ATC codes.}
\usage{
distance_between_codes(code1, code2, punishments)
}
\arguments{
\item{code1:}{string}

\item{code2:}{string}

\item{punishments:}{numeric vector}
}
\value{
numeric distance between two codes
}
\description{
The distance is calculated by given punishments.
For the same elements the distance is 0.
If either one of the codes is missing, the distance is 1.
}
\examples{
distance_between_codes("R03BA02", "R04BK10", c(5,4,3,2,1)) => 7
distance_between_codes("R03BA02", "R03BA02", c(5,4,3,2,1)) => 0
distance_between_codes("-1", "R03BA02", c(5,4,3,2,1)) => 1
}
