% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_additional_clusters_data}
\alias{get_additional_clusters_data}
\title{Find additional data about clusters.}
\usage{
get_additional_clusters_data(data, clusters, patients_data)
}
\arguments{
\item{data:}{dataframe}

\item{clusters:}{vector with lists for every cluster}
}
\value{
dataframe with additional data about every cluster
}
\description{
Find additional data about clusters.
}
