% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_distances}
\alias{get_distances}
\title{Gets values from certain indexes in matrix.}
\usage{
get_distances(matrix, indexes)
}
\arguments{
\item{matrix:}{matrix}

\item{indexes:}{numeric vector}
}
\value{
vector with matrix values where indexes intersect without duplicates
}
\description{
Expects for matrix to be symmetrical and removes duplicates.
}
