% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_drug_count}
\alias{get_drug_count}
\title{Get the count of a drug from data.}
\usage{
get_drug_count(data, drug)
}
\arguments{
\item{data:}{dataframe of drug codes with their frequencies}

\item{drug:}{string}
}
\value{
count of the drug in data, 0 if drug is not in data
}
\description{
Get the count of a drug from data.
}
