% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_filtered_patients}
\alias{get_filtered_patients}
\title{Get IDs of patients whose records count is in range.}
\usage{
get_filtered_patients(data, records_range)
}
\arguments{
\item{data:}{dataframe}

\item{records_range:}{numeric vector with two values - min and max record count}
}
\value{
vector that contains IDs of patients that fit into given range
}
\description{
Get IDs of patients whose records count is in range.
}
