% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_most_frequent_drug}
\alias{get_most_frequent_drug}
\title{Get the most frequent drug in the cluster.}
\usage{
get_most_frequent_drug(data, cluster)
}
\arguments{
\item{data:}{dataframe}

\item{cluster:}{numeric vector which has patients IDs}
}
\value{
string of most frequent drug code in given cluster
}
\description{
The most frequent drug per patient is taken into account.
}
