% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_most_frequent_values}
\alias{get_most_frequent_values}
\title{Get k most frequent values in a column.}
\usage{
get_most_frequent_values(data, column, k)
}
\arguments{
\item{data:}{dataframe}

\item{column:}{string}

\item{k:}{integer, k > 0}
}
\value{
vector containing k most frequent values in a column
}
\description{
Get k most frequent values in a column.
}
