% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_patient_drug_trajectory}
\alias{get_patient_drug_trajectory}
\title{Get patients drug trajectory of length n.}
\usage{
get_patient_drug_trajectory(data, id, n)
}
\arguments{
\item{data:}{dataframe}

\item{id:}{integer}

\item{n:}{integer n > 0}
}
\value{
vector that is patients drug trajectory that contains drug codes.
}
\description{
Get patients drug trajectory of length n.
}
