% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_sd_in_clusters}
\alias{get_sd_in_clusters}
\title{Gets average standard deviations in clusters if there are 1 to n clusters.}
\usage{
get_sd_in_clusters(matrix, max_clusters = 10)
}
\arguments{
\item{matrix:}{matrix}

\item{max_cluster:}{integer n > 0, maximum number of clusters}
}
\value{
vector with average standard deviations in clusters. In index 1 is average SD with 1 cluster, in index 2 with 2 clusters and so on.
}
\description{
Gets average standard deviations in clusters if there are 1 to n clusters.
}
