% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normalize}
\alias{normalize}
\title{Normalize all non-zero values in matrix with log10.}
\usage{
normalize(values)
}
\arguments{
\item{values:}{numeric matrix}
}
\value{
matrix where all non-zero values have been normalized.
}
\description{
Normalize all non-zero values in matrix with log10.
}
