% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_clusters_drugs_count}
\alias{plot_clusters_drugs_count}
\title{Create a plot which has a plot for each cluster that showcases occurrences of different drugs per patient.}
\usage{
plot_clusters_drugs_count(data, clusters, drugs, norm = FALSE)
}
\arguments{
\item{data:}{dataframe}

\item{clusters:}{vector of lists, every list represents one cluster with IDs}

\item{drugs:}{string vector which contains drug codes}
}
\value{
plot of drug occurrences in clusters
}
\description{
Create a plot which has a plot for each cluster that showcases occurrences of different drugs per patient.
}
