% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_elbow}
\alias{plot_elbow}
\title{Plots the elbow for matrix.}
\usage{
plot_elbow(matrix)
}
\arguments{
\item{matrix:}{matrix}
}
\value{
plot which shows average standard deviation in clusters if matrix is clustered into 1, 2, 3, ... clusters
}
\description{
Plots the elbow for matrix.
}
