% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_consecutive_elements}
\alias{remove_consecutive_elements}
\title{Remove consecutive elements of the same value.}
\usage{
remove_consecutive_elements(elements)
}
\arguments{
\item{elements:}{vector}
}
\value{
vector without multiple same elements in a row
}
\description{
Remove consecutive elements of the same value.
}
\examples{
remove_consecutive_elements(c(1, 2, 1, 1, 3)) => c(1, 2, 1, 3)
}
