% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{split}
\alias{split}
\title{Split string according to indexes.}
\usage{
split(string, indexes)
}
\arguments{
\item{string:}{string}

\item{indexes:}{numeric vector}
}
\value{
vector that contains pieces of the string
}
\description{
Split string according to indexes.
}
\examples{
split_into_pieces("abcdef", c(1, 2, 5)) => c("a", "bcd", "ef")
}
